#1779: fixed bug in emit function

+++ nes-runtime/src/Execution/Operators/Emit.cpp
@@ -44,7 +44,7 @@ void Emit::execute(ExecutionContext& ctx, Record& recordBuffer) const {
         ctx.emitBuffer(resultBuffer);
         auto resultBufferRef = ctx.allocateBuffer();
         emitState->resultBuffer = RecordBuffer(resultBufferRef);
-        emitState->outputIndex = 0ul;
+        emitState->outputIndex = (uint64_t) 0;
     }
 }
